# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /api/v1/approval_steps:
    get:
      tags:
        - ApprovalSteps
      summary: List ApprovalSteps
      description: Get a paginated list of ApprovalSteps with optional filtering and sorting
      operationId: listApprovalSteps
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApprovalStepList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - ApprovalSteps
      summary: Create ApprovalStep
      description: Create a new ApprovalStep
      operationId: createApprovalStep
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateApprovalStepRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApprovalStep'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/approval_step_templates:
    get:
      tags:
        - ApprovalStepTemplates
      summary: List ApprovalStepTemplates
      description: Get a paginated list of ApprovalStepTemplates with optional filtering and sorting
      operationId: listApprovalStepTemplates
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApprovalStepTemplateList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - ApprovalStepTemplates
      summary: Create ApprovalStepTemplate
      description: Create a new ApprovalStepTemplate
      operationId: createApprovalStepTemplate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateApprovalStepTemplateRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApprovalStepTemplate'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
...
    ApprovalStep:
      type: object
      required:
        - id
        - company_id
        - request_id
        - step_no
        - approver_kind
        - assigned_to
        - status
        - metadata
      properties:
        id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        request_id:
          type: string
          format: uuid
        step_no:
          type: integer
          format: int32
        approver_kind:
          type: string
        approver_ref:
          nullable: true
          type: string
          format: uuid
        assigned_to:
          type: string
          format: uuid
        delegated_from:
          nullable: true
          type: string
          format: uuid
        status:
          type: string
        acted_at:
          nullable: true
          type: string
          format: date-time
        comment:
          nullable: true
          type: string
        sla_due_at:
          nullable: true
          type: string
          format: date-time
        metadata:
          type: object
...
    CreateApprovalStepRequest:
      type: object
      required:
        - company_id
        - request_id
        - step_no
        - approver_kind
        - assigned_to
        - status
        - metadata
      properties:
        company_id:
          type: string
          format: uuid
        request_id:
          type: string
          format: uuid
        step_no:
          type: integer
          format: int32
        approver_kind:
          type: string
        approver_ref:
          nullable: true
          type: string
          format: uuid
        assigned_to:
          type: string
          format: uuid
        delegated_from:
          nullable: true
          type: string
          format: uuid
        status:
          type: string
        acted_at:
          nullable: true
          type: string
          format: date-time
        comment:
          nullable: true
          type: string
        sla_due_at:
          nullable: true
          type: string
          format: date-time
        metadata:
          type: object
...
    ApprovalStepList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ApprovalStep'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    ApprovalStepTemplate:
      type: object
      required:
        - id
        - company_id
        - policy_id
        - step_no
        - approver_kind
        - metadata
      properties:
        id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        policy_id:
          type: string
          format: uuid
        step_no:
          type: integer
          format: int32
        approver_kind:
          type: string
        approver_ref:
          nullable: true
          type: string
          format: uuid
        sla_hours:
          nullable: true
          type: integer
          format: int32
        all_of:
          nullable: true
          type: object
        metadata:
          type: object
...
    CreateApprovalStepTemplateRequest:
      type: object
      required:
        - company_id
        - policy_id
        - step_no
        - approver_kind
        - metadata
      properties:
        company_id:
          type: string
          format: uuid
        policy_id:
          type: string
          format: uuid
        step_no:
          type: integer
          format: int32
        approver_kind:
          type: string
        approver_ref:
          nullable: true
          type: string
          format: uuid
        sla_hours:
          nullable: true
          type: integer
          format: int32
        all_of:
          nullable: true
          type: object
        metadata:
          type: object
...
    ApprovalStepTemplateList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ApprovalStepTemplate'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []